try ap_int and sinf, cosf.

+++ pulsar_devkit/simple_examples/HLSIPs/METhls/MET.cpp
@@ -17,11 +17,11 @@ void MET_hw( ap_int<8> allPT[TotalN], ap_int<8> &missPT, ap_int<8> allPhi[TotalN
 	double totalY = 0;
 
 	for( i = 0; i < TotalN; i++){
-		totalX -= allPT[i]*cos(allPhi[i]);
-		totalY -= allPT[i]*sin(allPhi[i]);
+		totalX -= allPT[i]*cosf(allPhi[i]);
+		totalY -= allPT[i]*sinf(allPhi[i]);
 	}
 
-	missPT = sqrt(pow(totalX,2)+pow(totalY,2));
-	missPhi = acos( double( totalX / missPT ) );
+	missPT = sqrtf(pow(totalX,2)+pow(totalY,2));
+	missPhi = acosf( double( totalX / missPT ) );
 
 }
